Refactor merge postactions
