+++ code/test/CoreTemplateStudio.Core.Test/TestData/Merge/Source_postaction.cs
@@ -3,7 +3,6 @@
 //**}
 
 using Windows.UI.Xaml.Navigation;
-
 //{[{
 //USING
 //}]}
